fix a few more "extra" compiler warnings
Add casts for some int/long returns and assignments, e.g. i = (int)strlen(...)
Fix warning about type-punned pointer to ext_blocker_data.

+++ scst/include/backport.h
@@ -141,7 +141,7 @@ static inline unsigned int cpumask_next(int n, const cpumask_t *srcp)
 	/* -1 is a legal arg here. */
 	if (n != -1)
 		cpumask_check(n);
-	return find_next_bit(cpumask_bits(srcp), nr_cpumask_bits, n+1);
+	return (int)find_next_bit(cpumask_bits(srcp), nr_cpumask_bits, n+1);
 }
 
 /**
